+++ openyich-framework/src/main/java/com/openyich/framework/boot/config/JacksonConfiguration.java
@@ -1,5 +1,6 @@
 package com.openyich.framework.boot.config;
 
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.zalando.problem.ProblemModule;
@@ -7,13 +8,17 @@ import org.zalando.problem.validation.ConstraintViolationProblemModule;
 
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import com.fasterxml.jackson.module.afterburner.AfterburnerModule;
 
-/**
- * Support for Java date and time API.
- */
 @Configuration
+@ConditionalOnWebApplication
 public class JacksonConfiguration {
 
+  /**
+   * Support for Java date and time API.
+   * 
+   * @return the corresponding Jackson module.
+   */
   @Bean
   public JavaTimeModule javaTimeModule() {
     return new JavaTimeModule();
@@ -24,14 +29,27 @@ public class JacksonConfiguration {
     return new Jdk8Module();
   }
 
+  /**
+   * Jackson Afterburner module to speed up serialization/deserialization.
+   */
   @Bean
-  public ProblemModule problemModule() {
+  public AfterburnerModule afterburnerModule() {
+    return new AfterburnerModule();
+  }
+
+  /**
+   * Module for serialization/deserialization of RFC7807 Problem.
+   */
+  @Bean
+  ProblemModule problemModule() {
     return new ProblemModule();
   }
 
+  /**
+   * Module for serialization/deserialization of ConstraintViolationProblem.
+   */
   @Bean
-  public ConstraintViolationProblemModule constraintViolationProblemModule() {
+  ConstraintViolationProblemModule constraintViolationProblemModule() {
     return new ConstraintViolationProblemModule();
   }
-
 }